requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pip==22.3.1
polling==0.3.2
requests==2.31.0
setuptools==65.5.1
urllib3==2.0.3
wheel==0.38.4
python-telegram-bot==13.7
